+++ app/src/main/res/layout/activity_move_to_login.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:gravity="center"
-    tools:context=".MoveToLogin">
+    tools:context=".ui.login.MoveToLogin">
 
     <ImageView
         android:id="@+id/blurView"
